Configure pools.yaml inside Designate role
This adds the ability for a user to configure the Designate
pools.yaml file inside of the role by specifying an attribute.

Because the data required is yaml, it's a nice mapping to specify
the yaml attribute and have it dumped directly to the pools.yaml
file.

This allows users to use attributes from other plays (perhaps setting
up some complex DNS infrastructure in their cloud) and insert them
into Designate without having to write their own template or supply
their own file.

This also invokes the `designate-manage` command to load the pools.yaml
file into the Designate database, and simplifies the tests that
do the pools.yaml needful.

Change-Id: I11a849898bf33aa6b8aa6605296ac7fd733d7c01