Add venv_tag local fact
In order to make it easier to detect the currently deployed
venv for a service, and therefore allow smarter decisions
for things like upgrading, we implement the venv tag as a
local fact.

The file used to store facts will be the same for all
OpenStack services, with each service using its own section.

Example:

"ansible_local": {
    "openstack_ansible": {
        "designate": {
            "venv_tag": "14.2.1"
        }
    }
}

Change-Id: I6e0850cad13d129f8f781a754ee9481ae5bd9ab3