Use command instead of debug for handlers
Due to the debug message plugin the handler restart
messages show at the end of the playbook execution
which is a little confusing. Using debug also
requires setting changed_when to true which is a
little extra bit of code which we do not have to
carry.

Instead we use the command module which is simple,
works and less wordy.

Change-Id: Ic80a23df0d86258ae7931ebb0dea2bcea97187a2
(cherry picked from commit 9315d428c3637f53ff8e40bc82493246baa23828)