Allow the developer constraints to be adjusted
In order to do a developer mode that allows installation of packages
from local file, which take precedence over the developer_mode
constraints, we need to allow the order of the constraints to be changed.

This patch adds a "pip_install_developer_constraints" var which is used
to set the developer mode constraints. By default this will leave the
same behaviour but will allow additional constraints to be added, or the
developermode constraints file to be overriden altogether.

Additionally, fix bug where keystone_developer_mode variable was
incorrectly used for designate role.

Change-Id: I3d3ac65534dd71b6e5a956cc572a415e28e722fb